cleanup to use da_address
